Updated readme.md to prompt the user to run CameraControls.install()
I ran into problems the first time I used camera-controls, because I didn't copy the example usage code, which contains the line CameraControls.install( { THREE: THREE } ). This is neccesary for camera-controls to work, but is not explicitly stated. I've added a warning to remind the user about the importance of running that line.